requirements.txt
Django==1.9.4
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
Markdown==2.6.6
